# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API untuk Tubes Pemrograman",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tool/color-codes": {
            "get": {
                "tags": [
                    "ToolColorCode"
                ],
                "summary": "List tool color codes",
                "description": "Display a listing of the resource.",
                "operationId": "b13fb398fdcc533c42999ea9818ef592",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ToolColorCode"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tool-products": {
            "get": {
                "tags": [
                    "ToolProduct"
                ],
                "summary": "List tool products",
                "description": "Display a listing of the resource.",
                "operationId": "a4018f9f4c40196d873838c2b83a895a",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ToolProduct"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ToolProduct"
                ],
                "summary": "Create tool material",
                "description": "Store a newly created resource in storage.",
                "operationId": "18e08f5d4a875ed91ccc266d14a5b94a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreToolProductRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreToolProductRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreToolProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolProduct"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tool-products/{toolProduct}": {
            "get": {
                "tags": [
                    "ToolProduct"
                ],
                "summary": "Show tool material",
                "description": "Display the specified resource.",
                "operationId": "b197d0aa9ad34b1da5d02e55655d5ed1",
                "parameters": [
                    {
                        "name": "toolProduct",
                        "in": "path",
                        "description": "The tool material id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ToolProduct/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ToolProduct"
                ],
                "summary": "Update tool material",
                "description": "Update the specified resource in storage.",
                "operationId": "5cfb47f96b76c3b6d4aec11087d21cee",
                "parameters": [
                    {
                        "name": "toolProduct",
                        "in": "path",
                        "description": "The tool material id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ToolProduct/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateToolProductRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateToolProductRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateToolProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolProduct"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ToolProduct"
                ],
                "summary": "Delete tool material",
                "description": "Remove the specified resource from storage.",
                "operationId": "17f9a1622c200bc946a3bdcb747e3222",
                "parameters": [
                    {
                        "name": "toolProduct",
                        "in": "path",
                        "description": "The tool material id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ToolProduct/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModelNotFoundException": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "No query results for model [App\\Models\\ToolMaterial] 1"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The name field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UnprocessableContentException": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Wrong username or password"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The name field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreToolProductRequest": {
                "required": [
                    "tool_material_id",
                    "code",
                    "min_cutting_speed",
                    "max_cutting_speed"
                ],
                "properties": {
                    "tool_material_id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Tool product code",
                        "type": "string",
                        "example": "TP-001"
                    },
                    "name": {
                        "description": "Tool product name",
                        "type": "string",
                        "example": "Karbida"
                    },
                    "min_cutting_speed": {
                        "description": "Minimum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    },
                    "max_cutting_speed": {
                        "description": "Maximum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "UpdateToolProductRequest": {
                "required": [
                    "tool_material_id",
                    "code",
                    "min_cutting_speed",
                    "max_cutting_speed"
                ],
                "properties": {
                    "tool_material_id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Tool product code",
                        "type": "string",
                        "example": "TP-001"
                    },
                    "name": {
                        "description": "Tool product name",
                        "type": "string",
                        "example": "Karbida"
                    },
                    "min_cutting_speed": {
                        "description": "Minimum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    },
                    "max_cutting_speed": {
                        "description": "Maximum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "ToolColorCode": {
                "required": [
                    "code",
                    "color"
                ],
                "properties": {
                    "id": {
                        "description": "Tool color code ID",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Tool color code",
                        "type": "string",
                        "example": "R"
                    },
                    "color": {
                        "description": "Tool color code color",
                        "type": "string",
                        "example": "#FF0000"
                    },
                    "created_at": {
                        "description": "Tool color code created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Tool color code updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-01 00:00:00"
                    }
                },
                "type": "object"
            },
            "ToolItem": {
                "required": [
                    "tool_product_toolbox_id",
                    "tool_color_code_id"
                ],
                "properties": {
                    "id": {
                        "description": "Tool item ID",
                        "type": "integer",
                        "example": 1
                    },
                    "tool_product_toolbox_id": {
                        "description": "Tool product toolbox ID",
                        "type": "integer",
                        "example": 1
                    },
                    "tool_color_code_id": {
                        "description": "Tool color code ID",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Tool item created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Tool item updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-01 00:00:00"
                    },
                    "tool_product_toolbox": {
                        "$ref": "#/components/schemas/ToolProductToolbox"
                    },
                    "tool_color_code": {
                        "$ref": "#/components/schemas/ToolColorCode"
                    }
                },
                "type": "object"
            },
            "ToolMaterial": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Tool material ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Tool material name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "description": {
                        "description": "Description of the tool material",
                        "type": "string",
                        "example": "Karbida is a material used for tools"
                    },
                    "created_at": {
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "products": {
                        "description": "Tool products",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ToolProduct"
                        }
                    }
                },
                "type": "object"
            },
            "ToolProduct": {
                "required": [
                    "code",
                    "min_cutting_speed",
                    "max_cutting_speed"
                ],
                "properties": {
                    "id": {
                        "description": "Tool product ID",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Tool product code",
                        "type": "string",
                        "example": "TP-001"
                    },
                    "name": {
                        "description": "Tool product name",
                        "type": "string",
                        "example": "Karbida"
                    },
                    "min_cutting_speed": {
                        "description": "Minimum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    },
                    "max_cutting_speed": {
                        "description": "Maximum cutting speed (m/min)",
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "material": {
                        "description": "Tool product materials",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ToolMaterial"
                        }
                    },
                    "toolboxes": {
                        "description": "Tool product toolboxes",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ToolProductToolbox"
                        }
                    }
                },
                "type": "object"
            },
            "ToolProductToolbox": {
                "required": [
                    "tool_product_id",
                    "code"
                ],
                "properties": {
                    "id": {
                        "description": "Tool product toolbox ID",
                        "type": "integer",
                        "example": 1
                    },
                    "tool_product_id": {
                        "description": "Tool product ID",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Tool product toolbox code",
                        "type": "string",
                        "example": "TP-001"
                    },
                    "created_at": {
                        "description": "Timestamp of creation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp of last update",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01 00:00:00"
                    },
                    "tool_product": {
                        "$ref": "#/components/schemas/ToolProduct"
                    },
                    "tool_items": {
                        "description": "Tool items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ToolItem"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}